% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{calc_density}
\alias{calc_density}
\title{Calculate density function for Gaussian mixture model}
\usage{
calc_density(y, mu, sigma_sq, class_prob)
}
\arguments{
\item{y}{vector of points where density should be evaluated}

\item{mu}{vector of GMM means}

\item{sigma_sq}{vector of GMM variances}

\item{class_prob}{vector of GMM mixture probabilities}
}
\value{

}
\description{
Calculate density function for Gaussian mixture model
}
\examples{
# Simulate Data
n <- 1000
K <- 3
class_prob <- c(0.5, 0.3, 0.2)
mu <- rnorm(3)
sigma_sq <- 1/rgamma(3, 1,1)
class_label <- sample(1:3, n, replace = T, prob = class_prob)
y <- purrr::map_dbl(1:n, ~{
  rnorm(1, mu[class_label[.x]], sd = sqrt(sigma_sq[class_label[.x]]))
})
# Initialize Parameters
init <- list(mu = runif(3),
             sigma_sq = runif(3),
             class_prob = class_prob)
# Run EM
fit <- gmm_em(y, init)

ygrid <- seq(min(y), max(y), l = n)
den <- calc_density(ygrid, fit$mu, fit$sigma_sq, fit$class_prob)

# Plot result
df <- data.frame(y =y, class_label = as.factor(class_label))
ggplot(df) +
  geom_density(aes(x = y, color = class_label, group = class_label)) +
  geom_density(aes(x = y)) +
  geom_line(data = data.frame(x = ygrid, y = den), aes(x = x, y = y),
            color = "black", linetype = 2)
}
