% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleGEE_c.R
\name{HurdleGEE_c}
\alias{HurdleGEE_c}
\title{Generalized Estimating Equations, Truncated Count component of Longitudinal Hurdle Model}
\usage{
HurdleGEE_c(y, subjectID, N, X_c)
}
\arguments{
\item{y}{The vector of response counts, ordered by subject, time within subject.}

\item{subjectID}{The vector of subject ID values for each response.}

\item{N}{The number of subjects.}

\item{X_c}{The design matrix for all covariates in the truncated count component of the model, including an intercept.}
}
\description{
This function calculates the Generalized Estimating Equations (GEE) parameter estimates and standard errors for the truncated count component of a hurdle model for longitudinal excess zero count responses with independent working correlation structure, based on Dobbie and Welsh (2001).  Within the estimation method, responses are filtered to include only the positive counts from the original data.  Data must be organized by subject, and an intercept term is assumed for both the logistic and truncated count components of the model.  The function outputs a list with parameter estimates betaHat and parameter covariance estimate covEst, along with estimated rates muHat and residuals r_c.
}
\examples{
HurdleGEE_c()
}
\keyword{GEE}
