% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleGEE_l.R
\name{HurdleGEE_l}
\alias{HurdleGEE_l}
\title{Generalized Estimating Equations, Logistic component of Longitudinal Hurdle Model}
\usage{
HurdleGEE_l(y, subjectID, N, X_l)
}
\arguments{
\item{y}{The vector of response counts, ordered by subject, time within subject.}

\item{subjectID}{The vector of subject ID values for each response.}

\item{N}{The number of subjects.}

\item{X_l}{The design matrix for all covariates in the logistic component of the model, including an intercept.}
}
\description{
This function calculates the Generalized Estimating Equations (GEE) parameter estimates and standard errors for the logistic component of a hurdle model for longitudinal excess zero count responses with independent working correlation structure, based on Dobbie and Welsh (2001).  Responses are treated as binary indicators of 0 count values, therefore the probability modeled is that of a zero count.  Data must be organized by subject, and an intercept term is assumed for both the logistic and truncated count components of the model.  The function outputs a list with parameter estimates betaHat and parameter covariance estimate covEst, along with estimated probabilities pHat and residuals r_l.
}
\examples{
HurdleGEE_l()
}
\keyword{GEE}
