% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pACT.R
\name{pACT}
\alias{pACT}
\title{Actual Adjusted Minimum p-Value}
\usage{
pACT(pvals)
}
\arguments{
\item{pvals}{The vector of p-values to be adjusted.}
}
\description{
This function takes a list of ordered p-values and returns an adjusted minimum p-value.
}
\examples{
pACT()
}
\keyword{Multiple}
\keyword{Testing}
