% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pAdjusted.R
\name{pAdjusted}
\alias{pAdjusted}
\title{Multiple Testing p-Value Adjustment}
\usage{
pAdjusted(pvals)
}
\arguments{
\item{pvals}{The vector of p-values to be adjusted.}
}
\description{
This function adjusts a list of p-values for multiple testing.
}
\examples{
pAdjusted()
}
\keyword{Multiple}
\keyword{Testing}
