% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Representation graphique des clusters avec les donnees}
\usage{
plot_cluster(data, clusters)
}
\arguments{
\item{data}{data matrix}

\item{clusters}{numeric vector}
}
\description{
Representation graphique des clusters avec les donnees
}
\examples{
#data = data("iris")
#fit <- modelMixte(data,k)
#plot_cluster(data,fit$cluster)

}
