% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BIC.R
\name{plotBIC}
\alias{plotBIC}
\alias{plot_BIC}
\title{Representation graphique du BIC pour les clusters}
\usage{
plot_BIC(data, k)
}
\arguments{
\item{data}{data matrix}

\item{k}{number of clusters}
}
\description{
Representation graphique du BIC pour les clusters
}
\examples{
#k=3
#data = data("heterodata")
#plot_BIC(data,k)

}
