% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSGMM_Ber.R
\name{TSGMM_Ber}
\alias{TSGMM_Ber}
\title{Two-Step Generalized Method of Moments, Longitudinal Binary Outcome}
\usage{
TSGMM_Ber(yvec, subjectID, Zmat, Xmat, Tvec, N, mc = "EC",
  covTypeVec = c(-1))
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectID}{The vector of subject ID values for each response.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{Tvec}{The vector of times for each subject.}

\item{N}{The number of subjects.}

\item{mc}{The method of identifying appropriate moment conditions, either 'EC' for extended classification (default) or 'Types' for user-identified types.}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate, according to the order of the columns of Xmat.}
}
\description{
This function calculates the Generalized Method of Moments (GMM) parameter estimates and standard errors for longitudinal binary (0/1) responses.  This is modeled similarly to a Logistic Regression for Bernoulli responses.  The function allows for unbalanced data, meaning subjects can have different numbers of times of observation.  Both time-independent covariates and time-dependent covariates can be accommodated.  Time-dependent covariates can be handled either by specifying the type of each time-dependent covariate, or by allowing the data to determine appropriate moment conditions through the extended classification method.
}
\examples{
TSGMM_Ber()
}
\keyword{GMM}
