% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSGMM_Hurdle.R
\name{TSGMM_Hurdle}
\alias{TSGMM_Hurdle}
\title{Two-Step Generalized Method of Moments, Longitudinal Hurdle Model}
\usage{
TSGMM_Hurdle(y, subjectID, N, Tvec, X_l, Z_l, X_c, Z_c, mc = "EC",
  covTypeVec_l = c(-1), covTypeVec_c = c(-1))
}
\arguments{
\item{y}{The vector of response counts, ordered by subject, time within subject.}

\item{subjectID}{The vector of subject ID values for each response.}

\item{N}{The number of subjects.}

\item{Tvec}{The vector of times for each subject (N x 1)}

\item{X_l}{The design matrix for time-dependent covariates in the logistic component of the model, excluding an intercept.}

\item{Z_l}{The design matrix for time-independent covariates in the logistic component of the model, excluding an intercept.}

\item{X_c}{The design matrix for time-dependent covariates in the truncated count component of the model, excluding an intercept.}

\item{Z_c}{The design matrix for time-independent covariates in the truncated count component of the model, excluding an intercept.}

\item{mc}{The method of identifying appropriate moment conditions, either 'EC' for extended classification (default) or 'Types' for user-identified types.}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate, according to the order of the columns of X_l and X_c.}
}
\description{
This function calculates the Generalized Method of Moments (GMM) parameter estimates and standard errors for longitudinal excess zero count responses using a hurdle model formulation.  The function allows for unbalanced data, meaning subjects can have different numbers of times of observation.  Both time-independent covariates and time-dependent covariates can be accommodated.  Time-dependent covariates can be handled either by specifying the type of each time-dependent covariate, or by allowing the data to determine appropriate moment conditions through the extended classification method.  Data must be organized by subject, and an intercept term is assumed for both the logistic and truncated count components of the model.  The function outputs a list with parameter estimates betaHat_l and betaHat_c for the logistic and count components, respectively, along with parameter covariance estimates covEst_l and covEst_c.
}
\examples{
TSGMM_Hurdle()
}
\keyword{GMM}
