% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSGMM_PP.R
\name{TSGMM_PP}
\alias{TSGMM_PP}
\title{Two-Step Generalized Method of Moments, Truncated Count Component of Longitudinal Hurdle Model}
\usage{
TSGMM_PP(yvec, subjectID, Zmat, Xmat, Tvec, N, mc = "EC",
  covTypeVec = c(-1), betaI = c(1), r_c = c(-1))
}
\arguments{
\item{subjectID}{The vector of subject ID values for each response ((sum(Tvec)) x 1).}

\item{Zmat}{The design matrix for time-independent covariates ((sum(Tvec)) x K0).}

\item{Xmat}{The design matrix for time-dependent covariates ((sum(Tvec)) x Ktv).}

\item{Tvec}{The vector of times for each subject.}

\item{N}{The number of subjects.}

\item{mc}{The method of identifying appropriate moment conditions, either 'EC' for extended classification (default) or 'Types' for user-identified types.}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate, according to the order of the columns of Xmat.}

\item{betaI}{The initial parameter estimates.
r_c The vector of residuals from hurdle GEE with independent working correlation structure.}

\item{y}{The vector of positive count responses.  This vector must be organized by subject, and by time within subject ((sum(Tvec)) x 1).}
}
\description{
This function calculates the Generalized Method of Moments (GMM) parameter estimates and standard errors for the zero-truncated count component ("positive Poisson") of a hurdle model for longitudinal excess zero count responses.  This is modeled similarly to a Positive Poisson Zero-Truncated Regression using a log link.  The function allows for unbalanced data, meaning subjects can have different numbers of times of observation.  Both time-independent covariates and time-dependent covariates can be accommodated.  Time-dependent covariates can be handled either by specifying the type of each time-dependent covariate, or by allowing the data to determine appropriate moment conditions through the extended classification method.  Data must be organized by subject, and an intercept term is assumed.  The function outputs a list with parameter estimates betaHat along with parameter covariance estimates covEst.
}
\examples{
TSGMM_PP()
}
\keyword{GMM}
