% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validComb_EC_Binom.R
\name{validComb_EC_Binom}
\alias{validComb_EC_Binom}
\title{Generalized Method of Moments Valid Moment Combinations for Longitudinal Count (number of events from n trials) Responses, using Extended Classification}
\usage{
validComb_EC_Binom(ymat, Zmat, Xmat, betaI, Tvec, alpha)
}
\arguments{
\item{ymat}{The matrix of responses, ordered by subject, time within subject.  The first column is the number of successes, the second the number of failures.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{Tvec}{The vector of times for each subject.}

\item{alpha}{The significance level for inclusion of moment conditions.}
}
\description{
This function calculates the values of valid moment combinations for two-step Generalized Method of Moments using the extended classification method, applied to longitudinal data with count (0-n) outcomes.  It is assumed that the count represents the number of events from n identical trials, and that n is equal for all subjects and times.  This is modeled similarly to a Logistic Regression for Binomial responses.  It allows for unbalanced longitudinal data, meaning subjects can be observed for different numbers of times.  The function returns a vector "types" indicating validity of different moments conditions.
}
\examples{
validComb_EC_Binom()
}
\keyword{GMM}
