% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCBer_EC.R
\name{validMCBer_EC}
\alias{validMCBer_EC}
\title{Generalized Method of Moments Valid Moment Combinations for one Subject of Longitudinal Bernoulli Responses, using Extended Classification}
\usage{
validMCBer_EC(yvec, subjectIndex, Zmat, Xmat, betaI, T, Tmax, Count, types)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{Count}{A vector of running counts of the number of valid moment conditions for all subjects.}

\item{types}{A matrix indicating valid moment conditions for all time-dependent covariate parameters.}
}
\description{
This function calculates the values of the valid moment conditions for a single subject in a longitudinal study with binary outcomes.  It allows for unbalanced data, and uses the extended classification method to determine validity of moment conditions.  The function return a vector of valid moment condition values for subject i, along with an updated count vector of the number of valid moment conditions.
}
\examples{
validMCBer_EC()
}
\keyword{GMM}
