% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCBer_Types.R
\name{validMCBer_Types}
\alias{validMCBer_Types}
\title{Generalized Method of Moments Valid Moment Combinations for Longitudinal Bernoulli Responses, User-Defined Types}
\usage{
validMCBer_Types(yvec, subjectIndex, Zmat, Xmat, covTypeVec, betaI, T, Tmax,
  Count)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{Count}{A vector of running counts of the number of valid moment conditions for all subjects.}
}
\description{
This function calculates the values of valid moment combinations for two-step Generalized Method of Moments with user-defined types of time-dependent covariates, applied to longitudinal data with binary outcomes.  It allows for unbalanced longitudinal data, meaning subjects can be observed for different numbers of times.  The function returns a vector "types" indicating validity of different moments conditions.
}
\examples{
validMCBer_Types()
}
\keyword{GMM}
