% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCBinom_EC.R
\name{validMCBinom_EC}
\alias{validMCBinom_EC}
\title{Generalized Method of Moments Valid Moment Combinations for one Subject of Longitudinal Count (number of events from n trials) Responses, using Extended Classification}
\usage{
validMCBinom_EC(ymat, subjectIndex, Zmat, Xmat, betaI, T, Tmax, Count, types)
}
\arguments{
\item{ymat}{The matrix of responses, ordered by subject, time within subject.  The first column is the number of successes, the second the number of failures.}

\item{subjectIndex}{The location of the first index of subject i responses within ymat.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{Count}{A vector of running counts of the number of valid moment conditions for all subjects.}

\item{types}{A matrix indicating valid moment conditions for all time-dependent covariate parameters.}
}
\description{
This function calculates the values of the valid moment conditions for a single subject in a longitudinal study with count (0-n) outcomes.  It is assumed that the count represents the number of events from n identical trials, and that n is equal for all subjects and times.  This is modeled similarly to a Logistic Regression for Binomial responses.  It allows for unbalanced data, and uses the extended classification method to determine validity of moment conditions.  The function returns a vector of valid moment condition values for subject i, along with an updated count vector of the number of valid moment conditions.
}
\examples{
validMCBinom_EC()
}
\keyword{GMM}
