% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCPoi_EC.R
\name{validMCPoi_EC}
\alias{validMCPoi_EC}
\title{Generalized Method of Moments Valid Moment Combinations for one Subject of Longitudinal Count (Poisson) Responses, using Extended Classification}
\usage{
validMCPoi_EC(yvec, subjectIndex, Zmat, Xmat, betaI, T, Tmax, Count, types)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates ((N*T) x K0).}

\item{Xmat}{The design matrix for time-dependent covariates ((N*T) x Ktv).}

\item{betaI}{The current or initial estimates of the model parameters (1+K0+Ktv x 1).}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{Count}{A vector of running counts of the number of valid moment conditions for all subjects.}

\item{types}{A matrix indicating valid moment conditions for all time-dependent covariate parameters.}
}
\description{
This function calculates the values of the valid moment conditions for a single subject in a longitudinal study with count (Poisson) outcomes.  It allows for unbalanced data, and uses the extended classification method to determine validity of moment conditions.  The function returns a vector of valid moment condition values for subject i, along with an updated count vector of the number of valid moment conditions.
}
\examples{
validMCPoi_EC()
}
\keyword{GMM}
