% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCPP_Types.R
\name{validMCPP_Types}
\alias{validMCPP_Types}
\title{Generalized Method of Moments Valid Moment Combinations for Longitudinal Zero-Truncated Count Responses, User-Defined Types}
\usage{
validMCPP_Types(yvec, subjectIndex, Zmat, Xmat, covTypeVec, betaI, T, Tmax,
  Count)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{Zmat}{The design matrix for time-independent covariates ((N*T) x K0).}

\item{Xmat}{The design matrix for time-dependent covariates ((N*T) x Ktv).}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate.}

\item{betaI}{The current or initial estimates of the model parameters (1+K0+Ktv x 1).}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{Count}{A vector of running counts of the number of valid moment conditions for all subjects.}
}
\description{
This function calculates the values of valid moment combinations for two-step Generalized Method of Moments with user-defined types of time-dependent covariates, applied to longitudinal data with zero-truncated count ("positive Poisson") outcomes.  It allows for unbalanced longitudinal data, meaning subjects can be observed for different numbers of times.  The function returns a vector "types" indicating validity of different moments conditions.
}
\examples{
validMCPP_Types()
}
\keyword{GMM}
