% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMDNor_EC.R
\name{validMDNor_EC}
\alias{validMDNor_EC}
\title{Generalized Method of Moments Valid Moment Combination Derivatives for one Subject of Longitudinal Continuous Responses, using Extended Classification}
\usage{
validMDNor_EC(yvec, subjectIndex, Zmat, Xmat, betaI, T, Tmax, types)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{types}{A matrix indicating valid moment conditions for all time-dependent covariate parameters.}

\item{N}{The number of subjects.}
}
\description{
This function calculates the values of the derivatives of all valid moment conditions for a single subject in a longitudinal study with continuous (normal) outcomes.  It allows for unbalanced data, and uses the extended classification method to determine validity of moment conditions.  The function returns a matrix of derivatives for all valid moment condition for subject i.
}
\examples{
validMDNor_EC()
}
\keyword{GMM}
