% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMDNor_Types.R
\name{validMDNor_Types}
\alias{validMDNor_Types}
\title{Generalized Method of Moments Valid Moment Combination Derivatives for one Subject of Longitudinal Normal Responses, User-Defined Types}
\usage{
validMDNor_Types(yvec, subjectIndex, Zmat, Xmat, covTypeVec, betaI, T, Tmax)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates.}

\item{Xmat}{The design matrix for time-dependent covariates.}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate.}

\item{betaI}{The current or initial estimates of the model parameters.}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{N}{The number of subjects.}
}
\description{
This function calculates the values of the derivatives of all valid moment conditions for a single subject in a longitudinal study with continuous (normal) outcomes.  It allows for unbalanced data, and uses user-defined types of time-dependent covariates to determine validity of moment conditions.  The function returns a matrix of derivatives for all valid moment condition for subject i.
}
\examples{
validMDNor_Types()
}
\keyword{GMM}
