% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMDPP_EC.R
\name{validMDPP_EC}
\alias{validMDPP_EC}
\title{Generalized Method of Moments Valid Moment Combination Derivatives for one Subject of Longitudinal Zero-Truncated Count Responses, using Extended Classification}
\usage{
validMDPP_EC(yvec, subjectIndex, Zmat, Xmat, betaI, T, Tmax, types)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates ((N*T) x K0).}

\item{Xmat}{The design matrix for time-dependent covariates ((N*T) x Ktv).}

\item{betaI}{The current or initial estimates of the model parameters (1+K0+Ktv x 1).}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{types}{A matrix indicating valid moment conditions for all time-dependent covariate parameters.}

\item{N}{The number of subjects.}
}
\description{
This function calculates the values of the derivatives of all valid moment conditions for a single subject in a longitudinal study with zero-truncated count ("positive Poisson") outcomes.  It allows for unbalanced data, and uses the extended classification method to determine validity of moment conditions.  The function returns a matrix of derivatives for all valid moment condition for subject i.
}
\examples{
validMDPP_EC()
}
\keyword{GMM}
