% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMDPP_Types.R
\name{validMDPP_Types}
\alias{validMDPP_Types}
\title{Generalized Method of Moments Valid Moment Combination Derivatives for one Subject of Longitudinal Zero-Truncated Count Responses, User-Defined Types}
\usage{
validMDPP_Types(yvec, subjectIndex, Zmat, Xmat, covTypeVec, betaI, T, Tmax)
}
\arguments{
\item{yvec}{The vector of responses, ordered by subject, time within subject.}

\item{subjectIndex}{The location of the first index of subject i responses within yvec.}

\item{Zmat}{The design matrix for time-independent covariates ((N*T) x K0).}

\item{Xmat}{The design matrix for time-dependent covariates ((N*T) x Ktv).}

\item{covTypeVec}{The vector indicating the type of each time-dependent covariate.}

\item{betaI}{The current or initial estimates of the model parameters (1+K0+Ktv x 1).}

\item{T}{The number of time points for subject i.}

\item{Tmax}{The maximum number of times of observation among all subjects.}

\item{N}{The number of subjects.}
}
\description{
This function calculates the values of the derivatives of all valid moment conditions for a single subject in a longitudinal study with zero-truncated count ("positive Poisson") outcomes.  It allows for unbalanced data, and uses user-defined types of time-dependent covariates to determine validity of moment conditions.  The function returns a matrix of derivatives for all valid moment condition for subject i.
}
\examples{
validMDPP_Types()
}
\keyword{GMM}
