\name{wage}
\docType{data}
\alias{wage}
\title{Labor Data}
\description{
Data used to measure return to education by Griliches (1976)
}
\usage{data(wage)}
\format{
A data frame containing 20 cross-sectional vectors. 
\describe{
  \item{AGE, AGE80}{Age in 1969 and 1980 respetively}
  \item{EXPR, EXPR80}{Working experience in 1969 and 1980 respetively} 
  \item{IQ}{IQ measure of the individual} 
  \item{KWW}{A test score} 
  \item{LW, LW80}{Log wage in 1969 and 1980 respectively} 
  \item{MED}{Mother education}
  \item{MRT, MRT80}{}
  \item{RNS, RNS80}{}
  \item{S, S80}{Schooling in 1969 and 1980 respetively}
  \item{SMSA, SMSA80}{} 
  \item{TENURE, TENURE80}{Tenure in 1969 and 1980 respetively} 
  \item{YEAR}{} 
}
}

\source{\url{http://fhayashi.fc2web.com/datasets.htm}}


\keyword{datasets}
