\name{Finance}
\docType{data}
\alias{Finance}
\title{Returns on selected stocks}
\description{
Daily returns on selected stocks, the Market portfolio and factors of Fama and French from 1993-01-05 to 2009-01-30 for CAPM and APT analysis 
}
\usage{data(Finance)}
\format{
A data frame containing 24 time series. Dates are reported as rownames(). In the following description, compagny symboles are used.
\describe{
\item{WMK}{Returns of WMK} 
\item{UIS}{Returns of UIS} 
\item{ORB}{Returns of ORB} 
\item{MAT}{Returns of MAT} 
\item{ABAX}{Returns of ABAX} 
\item{T}{Returns of T} 
\item{EMR}{Returns of EMR} 
\item{JCS}{Returns of JCS} 
\item{VOXX}{Returns of VOXX}
\item{ZOOM}{Returns of ZOOM} 
\item{TDW}{Returns of TDW} 
\item{ROG}{Returns of ROG} 
\item{GGG}{Returns of GGG} 
\item{PC}{Returns of PC} 
\item{GCO}{Returns of GCO} 
\item{EBF}{Returns of EBF} 
\item{F}{Returns of F} 
\item{FNM}{Returns of FNM} 
\item{NHP}{Returns of NHP} 
\item{AA}{Returns of AA} 
\item{rf}{Risk-free rate,}
\item{rm}{Return of the market portfolio of Fama-French,}
\item{hml}{Factor High-Minus-Low of Fama-French,}
\item{smb}{Factor Small-Minus-Big of Fama-French.}
}
}

\source{\url{http://ca.finance.yahoo.com/} and \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/}}

\keyword{datasets}
