\name{Finance}
\docType{data}
\alias{Finance}
\title{Returns on selected stocks}
\description{
Daily returns on selected stocks, the Market portfolio and factors of Fama and French from 1993-01-05 to 2009-01-30 for CAPM and APT analysis 
}
\usage{data(Finance)}
\format{
A data frame containing 24 time series. Dates are reported as rownames(). In the following description, company symboles are used.
\describe{
\item{WMK}{Returns of WEIS MARKETS INC} 
\item{UIS}{Returns of UNISYS CP NEW} 
\item{ORB}{Returns of ORBITAL SCIENCES CP} 
\item{MAT}{Returns of Mattel, Inc.} 
\item{ABAX}{Returns of ABAXIS, Inc.} 
\item{T}{Returns of AT&T INC.} 
\item{EMR}{Returns of EMERSON ELEC CO} 
\item{JCS}{Returns of Communications Systems Inc.} 
\item{VOXX}{Returns of Audiovox Corp.}
\item{ZOOM}{Returns of ZOOM Technologies Inc.} 
\item{TDW}{Returns of TIDEWATER INC} 
\item{ROG}{Returns of Rogers Corporation} 
\item{GGG}{Returns of Graco Inc.} 
\item{PC}{Returns of Panasonic Corporation} 
\item{GCO}{Returns of Genesco Inc.} 
\item{EBF}{Returns of ENNIS, INC} 
\item{F}{Returns of FORD MOTOR CO} 
\item{FNM}{Returns of FANNIE MAE} 
\item{NHP}{Returns of NATIONWIDE HLTH PROP} 
\item{AA}{Returns of ALCOA INC} 
\item{rf}{Risk-free rate of Fama-French}
\item{rm}{Return of the market portfolio of Fama-French}
\item{hml}{Factor High-Minus-Low of Fama-French}
\item{smb}{Factor Small-Minus-Big of Fama-French}
}
}

\source{Yahoo Finance and \url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/}}


\keyword{datasets}
